pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "arbiter"
version = "0.1.0"
description = "Universal quality measurement engine for AI agent output"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.23,<1.0"]

[project.scripts]
arbiter = "arbiter.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["src"]
